requirements.txt
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
streamlit==1.38.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.13
